reworked museum api to make less calls to the MET api
